requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
conda==4.3.16
cycler==0.10.0
Django==3.0.11
django-cors-headers==3.6.0
django-filebrowser==3.13.2
django-grappelli==2.14.3
django-multitype-file-field==0.0.5
djangorestframework==3.12.2
ez-setup==0.9
generics==3.0.0
idna==2.10
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.4
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.20.1
Pillow==8.0.1
pycosat==0.6.3
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
